Add rasterio mask method
An attempt to use the built in `rasterio.mask.mask` method to mask out
portional of a raster from the supplied geojson polygon.  This
implementation appears to increase execution time by about 50% over the
existing implementation.  I believe this is because the method actually
fills the raster with the `nodata` value, which then needs to be
converted to a numpy masked array in order for fast array
manipulation.

This commit if mostly for posterity and will remove in the future.