Add parallel elevation vector extraction
Allows to read in a raster, compute min/max elevation in a masked area
and then extract vector features from that corresponding elevation
increment. The processing is done in parallel up to the number of cpus
available and the raster is shared in-memory. Notably, the output is a
GeometryCollection and not a single MultiPart polygon.  The unioning is
especially cpu intesive and may not be necessary for the final output.